tle

Module stream_ciphers

Structs

AESGCMStreamCipherProvider

This provides the AES_GCM stream cipher, allowing message to be encrypted and decrypted under AES_GCM

AESOutput

The output of AES_GCM Encryption

Enums

Error

Errors that mayb be encountered with using a stream cipher

Traits

StreamCipherProvider

Something that provides encryption and decryption using a stream cipher